requirements.txt
fastapi
uvicorn
python-pptx
Pillow
requests
python-dotenv
pydantic
python-multipart
